pyproject.toml
[tool.poetry]
name = "bq-baseball"
version = "0.1.0"
description = "BigQuery Baseball Dataset"
authors = ["Shinichi Nakagawa <spirits.is.my.rader@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-bigquery = "^1.24.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
